requirements.txt
streamlit
pandas
sqlite3
streamlit_js_eval
mysql.connector